feat: use Context API
