+++ packages/svelte/package.json
@@ -3,12 +3,12 @@
 	"type": "module",
 	"version": "0.0.0",
 	"description": "Svelte utilities for anywidget",
-	"main": "index.js",
+	"main": "src/index.js",
 	"types": "dist/index.d.ts",
 	"exports": {
 		".": {
 			"types": "./dist/index.d.ts",
-			"import": "./index.js"
+			"import": "./src/index.js"
 		}
 	},
 	"scripts": {
