+++ packages/svelte/src/constants.js
@@ -1,0 +1,2 @@
+export const MODEL_CONTEXT_NAME = "__anywidget.model";
+export const STORES_CONTEXT_NAME = "__anywidget.stores";
